Ready to go
-Cleaned up the code and created the CarSelectedBox component who takes care of the details when clicking the reservation button. The details can be later added for the "final" product.

-Put all cars related code in the CARS folder for simplicity on the components folder.

-It's ready to go to push to the main branch.
