+++ client/components/CARS/carInfo.jsx
@@ -1,11 +1,11 @@
 import React from 'react';
 
-const CarInfo = ({ car }) => {
+const CarInfo = ({ car, onClick }) => {
   const { id, name, href, imageSrc, imageAlt, price } = car;
 
   return (
     <a key={id} href={href} className="group">
-      <div className="aspect-w-1 aspect-h-1 w-full overflow-hidden rounded-lg bg-gray-200 xl:aspect-h-8 xl:aspect-w-7">
+      <div className="aspect-w-1 aspect-h-1 w-full overflow-hidden rounded-lg bg-gray-200 xl:aspect-h-8 xl:aspect-w-7" onClick = {onClick}>
         <img
           src={imageSrc}
           alt={imageAlt}
