+++ client/pages/customer/customerView.jsx
@@ -5,7 +5,7 @@ import React, { useState, useEffect } from "react";
 import { useSession } from "next-auth/react";
 import AccessDenied from "../../components/accessDenied";
 import CustomerConsole from "../../components/CustomerConsole";   
-import CarDisplay from "../../components/carDisplay";
+import CarDisplay from "../../components/CARS/carDisplay";
 
 const inter = Inter({ subsets: ["latin"] });
