+++ client/pages/index.jsx
@@ -4,7 +4,7 @@ import Image from "next/image";
 import { Inter } from "next/font/google";
 import Header from "../components/Header";
 const inter = Inter({ subsets: ["latin"] });
-import CarDisplay from "../components/carDisplay";
+import CarDisplay from "../components/CARS/carDisplay";
 
 export default function Home() {
   return (
